requirements.txt
langchain
chromadb
openai
bs4
nest_asyncio
pandas
asyncio
gradio
faiss-cpu
langchain_experimental
playwright
tiktoken
transformers
asyncpg
google-cloud-bigquery
pandas
vertexai
langchain-google-genai
langchain-google-vertexai
gcsfs
jupyterlab
langchain-community
langchain-openai
psycopg2
asyncpg
langchainhub
db-dtypes
google-api-python-client